Move aws_stack.extract_access_key_id_from_auth_header
